requirements.txt
imaplib2
pymupdf
requests
Pillow
schedule